requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
bleach==6.1.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-rest-auth==5.0.2
Django==5.0
django-cachalot==2.6.2
django-celery-results==2.5.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
drf_api_logger==1.1.15
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
kombu==5.3.5
multidict==6.0.4
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.11.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
